Rendering Bunny Hole in the Sidebar
At this point, the Bunny Hole can be rendered in the sidebar and will
even respond to updates as the user navigates. However, some other kinks
need to be worked out before pushing this to the main branch.

In particular, we need to dynamically style the nodes to be indented
based on the nesting of the BunnyHole. We also need to figure out why so
many messages are being sent (I suspect useEffect shenanigans).