Version 0.0.7 -- GUI Progression
Okay, we've been away for a bit, but this change makes a lot of progress
on the GUI. Changes include buttons to delete nodes, interfacing to
reorder nodes with both a two-click process and a click-and-drag
process, a toolbar (no functionality yet), and some cute animations and
transitions.

Further commits will polish up this functionality, but it's good to have
the rudimentary implementation in place!