Node Edit Modal
This commit adds a modal that allows the user to edit a particular node
in the Bunny Hole. The modal disallows the user from interacting with
any BunnyHole elements while it's active (even if they try to hit the
Tab key), and is responsive to clicking outside the dialog box, pressing
Enter or Escape to submit or exit, and clicking the providing Confirm
and Cancel buttons.