Freeze/Unfreeze Bunny Hole + Webpack Improvements
This commit adds the ability to "freeze" the Bunny Hole, or to tell it
to stop listening for new webnav events. Of course, the Bunny Hole can
also be unfreezed.

This also improves my Webpack configuration file. Now the entry points
are unified into a single configuration file, button images are bundled,
and cleaning the dist folder should be more consistent for me.