Preliminary Options UI
This commit implements a basic UI for BunnyHole options and theme
options. However, limitations with the popup menu means that we're going
to have to move this UI to the sidebar, not the popup. In particular,
the popup menu will close whenever we open a color picker dialogue,
which will possibly sever the connection to the popup script and
certainly make for a frustrating user experience.

We also reorganize some of the file structure in this commit.

All in all, this is a substantial commit, but a lot more work is
necessary to get this up and running well.