Basic Themes
This commit allows the user to switch between two preset themes.
The dark mode theme needs to be reworked, since it uses red to confirm
and that's just confusing.
I think we can also modify themes to reuse main or accent colors for
tooltips instead of having their own set of colors.
But, for now, this is a satisfactory implementation.

Also note that we're missing an immediate switch in the sidebar if the
theme is changed in the popup window.